It is another day and I am ready to fight again.
As annoying as it is, I do not have enough to meet my goals. As a side effect of implementing the L2 RNG, I now have the matrix inverse functions ready for me in the library so I could implement the vanilla PRONG methods.

Unfortunately for me I want to make it iterative.

I started RL for real over two months ago, and I still seem to be in research mode.

I got most of the talk yesterday which is definitely a milestone in my understanding, but there are still some glaring holes in my knowledge. Regarding K-FAC though I can watch such a talk freely now, I still do not actually understand how the weight gradients could be from the expectations of the inputs and and errors.

The math in my head does not check out. Multiplying the two FIM approximations would get a dimensionality error.

I'll have to watch the relevants part of the talk, or better yet, read the paper again.

That is just for KFAC. For relative gradient, I still do not exactly understand the difference between `f(x + d x)` and `f (x + d)`. Cardoso never really explains what possessed him to try the former in place of the later. This is not what I remember from my calculus class.

Also related to this, I think I should finally learn mirror descent. I never went through the paper properly last week.

Also since all of a sudden source separation is relevant to NNs, I am going to have to study this as well.

The ultimate goal would be to implement BPRONG with fully relative gradient kind of updates.

Updating the whitening and centering parameters and then everything else would be unstable. It would be absolute perfection if I could get it all down perfectly.

I think at that point I'd have enough to satisfy even me. I'll be able to start work on the agents with the expectations of actually winning in real life.

Today, let me start with the K-FAC paper. I'll go through it again.