Let me try it again. Despite everything, I am still not suspecting my own code.
By now I have over 3 years of experience putting 8h per day into this art so my intuition has been sharpened to a high degree.

If the actual layer code is not wrong then that leaves something very simple. The cost function.

Of course, there is no way that square error that I am using is wrong. What I am suspecting is the the RNG is so brittle that even slight chance from the FIM method completely breaks it by making it degenerate into standard gradient descent.

Let me start by testing this hypothesis then. If it turns out to be right this will be the deathblow to the method.