I've confirmed it. The assumption I had that was broken is that SGD would be suitable for approximating the inversee.
If I initialize the matrix to really high random values, it will get it down, but not below 61.25 which is the baseline zero matrix.

The reason why I am not getting any better results than SGD is because it never moves away from it in the first place.

I'll try the whitening matrix updates from the EASS paper tomorrow.