So now I am roughly where I was yesterday. Let me stop reading `To Be a Power In The Shadows` and do some work.
The lecture I watched yesterday was really informative, but I need more.

https://www.youtube.com/playlist?list=PLIljB45xT85DWUiFYYGqJVtfnkUFWkKtP

Maybe my next goal will be to watch a few videos from this playlist. Ke Sun did give a suggestion to learn some differentiable geometry if I want to know more about the field.

I am done with the RFIM stuff completely - I no longer have an interest in it. I fulfilled my obligation towards it.

Beside that, it is time for a half yearly review.

Today and tomorrow, I am going to take the time to compose the update for the Simulacrum blog. That usually takes me a while.

Let me start it now then.