https://arxiv.org/abs/1612.05877 Deep Learning on Lie Groups for Skeleton-based Action Recognition
I want to learn about the Lie group related stuff because relative gradient is related to this. Also since EASS is confusing me, maybe I should watch a few videos on ICA and such?

I've been focusing on relative gradient, but the truth is I am not sure what EASS is supposed to be doing.

Let me start my day like this.