I understand the Woodbury matrix identity now. I suppose it is a viable choice if I want to reproject really often.
The ideal would be though to derive natural gradient style updates similar to the ones in the EASS paper for BPRONG.

Agh. There is no choice anymore.

Either I use the straight inverse, or I dedicate the next month to painstakingly studying information geometry in hopes that I manage to get enough insight to help me crack this.

By this point I've managed to go through the easy material and only those books and the papers are waiting for me.

It is a risky proposition to be sure, but I am just going to have to gamble here?

I am thinking back on my life. Did I ever really have programming talent? Or did I merely decide that I have to be good at programming and something in me responded to that wish? Programming was the only thing I ever put the weight of all my pride.

I am going to do the same for math. If I lack talent I am going to force it - one step at a time.

The plan will be to focus on Amari's book 30 pages at a time until the end of next month. I've never tried something like that. Maybe it will bring some result in the end.

For now I am done for the day. Time for fun.