I think I will start the day by going through Amari's book.
I cant' help it. I am going to have to familiarize myself with this material in any way that I can.

With the RUDDER paper it is amazing how quicky things are falling into place. I predicted a 1000x increase, but that thing is already here.

What is needed is that I figure out natural gradient methods for good, combine it with the stuff by Seyn and combine it with RUDDER.

Then I will not just be able to not just beat the micros, but pretty much scrape the gambling dens clean.

Today is kind of a special day for me, so I might be taking off later to have some fun outside, but I'll do what I can till then.