Done with chilling. Today my goal will be to make it through the rest of the book and then I will look at the slides for the information geometry PhD course. I'll also aim to go through Nihat Ay's book.
After that I need to make time for the recurrent K-FAC paper.

After that, come the directional derivatives. I want to see if they are used in NNs anywhere.

The worst part of all of this is how inefficient the whole learning process is. I have no choice, but to go through it.

I'll come out on top eventually. The half year update was disappointing, but hopefully by the end of the year I will have something to show for it. I just need to continue moving forward. At this point that is the most I can expect from myself.

My programming skills are top notch. If I can get my AI related skills to a similar level at that point I will be set. I will be able to effect real change in the world the doors will start opening for me.