Done with the books and the slides.
Let me take a break here. I need to gather my thoughts.

Both of the books and the slides are too much for me. I am really not getting much from simply following along.

Since the next thing on the agenda are directional derivatives, maybe to learn about them I should just go through the damn differential geometry course?

That will be the only choice if I want to progress from here. I do not get the sense that Amari's book is intended to be all that difficult, but despite that I cannot follow it all. Information geometry seems to be mostly differential geometry anyway so I do not think I will be needing to go for statistic courses when I am done.

Agh, damn it. Going through those 35 videos will be boring, but I might as well do it. I might as well add finishing the numerical optimization course and going through the convex optimization course to the mix. All in all, this will be enough to keep me busy for the next month.

...Uah. I am already yawning here just thinking about it. For the rest of the day, how about I go through two more videos?

That will be my goal for the rest of day.