I thought differential geometry is going to be like Amari's book, but in the lst 3 lectures it really picked up. I am into it now. Yeah, I am a bit bored, but it is at a comfortable level. I'll be done with it in a week. Originally I intended to slowly mix it with the rest, but now I want to go all the way to the end quickly.
I'll really be disappointed if I do not learn how to take advantage of directional derivatives by the end of it.

When he started talking about an algebraic approach it really made me realize that the lecturer and me are on the same wavelength.

Despite the recent stumbling around information geometry I did level up in the past few years. I just need to take it a step further and I will reach the finish line.