Lecture 9 was just sad. I am finding that I can understand algebra well enough, but I am simply lost when it comes to trying to understand what the guy is trying to explain visually.
It was same for that whole vertex business in the finite fields lecture.

The truth is, after the third dimensions, I am absolutely not sure what I would do with any of this understanding even if I had it. Neural nets have a massive number of dimensions.

Well, I'll do what I can. As per plan I'll finish this course and then try Amari's book again. If I get nothing from the course, I'll abandon the plan of trying to learn information geometry and try to at least figure out what the hell relative gradient updates are supposed to be doing and how directional derivatives fit into the whole scheme.

Well, since the lectures are so short and I only have one left before I clear 5, I think I will add two more to the mix today.

Even though it is boring, this course is not that difficult. I'll be done by the end of the month which is a relief.