Max Payne 3 is great. It is just as good as the previous two in the series. Right now I am just coming down from the high of finishing it on Old School difficulty. It was not too hard.
Fuuuu...

Though discovering the essence of self improvement has changed me and now I have a single minded obsession to increase my skills, I know that in the future if I ever feel that I am losing my connection to who I was, all I will need to do is play a game. Friends and family might be one's connection to humanity, but games are one's connection to the self.

I really do owe a bit to my recent visitor for whom I got it for.

Dark Souls will be next on my list.

I've earned this leisure, but as always, a part of me is telling me that I need to hurry. My lifespan will not be forever and the long term is quite a scary thing to behold. I feels like the last 2.5 months just vanished into thin air for me.

And I am going to have to sacrifice a bit more in a gambit too. Natural gradients are one of the rare things which I do not consider filler, so I need to upgrade my understanding. Once that is done, Zap-Q and Rudder are next. Those two are fundamental to fixing the grave flaws in the current RL algorithms. After that I'll essentially have everything I need to make smooth progress.