"9:10am. https://www.youtube.com/watch?v=ZFd0I2Ry5KA LM1. Directional Derivatives and Optimization with Constraints
What is this?

Let me watch this and I'll continue with the other lectures.

https://www.youtube.com/playlist?list=PL1nlTMCcBDWB5j-dkU5JuBDU6FV1W2wid

Here is the playlist with this.

9:30am. So in order to understand directional derivatives I need to understand constrained optimization.

Well, now I have a reason to go through the convex optimization course.

9:50am. https://www.youtube.com/watch?v=4NXz4tU_pH0&list=PLIljB45xT85DWUiFYYGqJVtfnkUFWkKtP&index=13

Let me stop here.

I think I am going to stop with the differential geometry course. I honestly can't be assed to learn all of this just to get the tiny bit that I actually need. Since I have lead on the directional derivatives, let me go through the stuff in that playlist. Also rather than mess with all of this, it might be better for me to look into signal processing.

...No forget it. I can't be lead astray any longer. I think it would be fine if I spread my efforts a little. Let me try that playlist. Then I'll give convex optimization a try."

The above is an excerpt from my journal.

The quest continues as usual. Let me go through a few of those lectures on the use of directional derivatives for optimization and then I will try figuring out the W*W^T rule by Amari directly.