"4:05pm. Is optimization on a manifold simply constrained optimization? Is natural gradient learning constrained optimization?
https://www.semanticscholar.org/paper/Learning-Manifolds-in-the-Wild-Hegde-Sankaranarayanan/83a2188ae1ab96039cb25112481116cdfb573138
Learning Manifolds in the Wild

I'll read this later.

https://pdfs.semanticscholar.org/b925/43d51c584f47a3e036842b545ee1513c9118.pdf
Adaptive On-Line Learning Algorithms for Blind Separation - Maximum Entropy and Minimum Mutual Information

https://pdfs.semanticscholar.org/52ff/21aabaecdffe8dff063d0bba113a6a432fac.pdf
Natural Gradient Descent for Training Multi-Layer Perceptrons

4:40pm. Damn, it is already this time. My focus is starting to weaken. I am not sure I will manage the playlist, but I will try my best.

I'll read the above papers tomorrow.

https://www.researchgate.net/publication/222957173_Optimization_on_Lie_manifolds_and_pattern_recognition
Optimization on Lie manifolds and pattern recognition

https://pdfs.semanticscholar.org/3454/2c75da4126f1bcaf84e66ac5f94d1e703955.pdf
Lie Group Methods for Optimization with Orthogonality Constraints

There is a bunch of stuff when I Google for 'lie optimization'.

"Optimization of a cost function J(W) under an orthogonality
constraint WWT = I is a common requirement for ICA methods.
In this paper, we will review the use of Lie group methods to perform
this constrained optimization. Instead of searching in the space of n × n
matrices W, we will introduce the concept of the Lie group SO(n) of orthogonal
matrices, and the corresponding Lie algebra so(n). Using so(n)
for our coordinates, we can multiplicatively update W by a rotation matrix
R so that W′ = RW always remains orthogonal. Steepest descent
and conjugate gradient algorithms can be used in this framework."

This paper seems promising. I'll leave it for later.

https://www.youtube.com/watch?v=bXM5PgAQvcA&list=PLo6qq_MJTOZ2ZHdr-aMdZNcsjkbGD65uR

I'll stop on this video.

I'll aim go through half of the rest of the playlist tomorrow.

5pm. I think I will simply play it smart here and stop for the day. I had enough.

Hopefully the Lie algebra series will go somewhere. The differential geometry course just wants to talk about math which gets boring quite quickly. Beyond a certain point the talent does come into it. I mentioned in the update that I was always much better than the classmates when it came to programming. A large part of it was my ability to create motivation which they apparently lacked. It might be the case that in the math regime I myself lack this crucial piece.

Well, up to a certain point I can hack it. Hopefully I will be able to grasp the essence of optimization over Lie groups.

After I am done with that I'll stop messing around with this and implement BPRONG using the Cholesky update in the CMA paper.

Well, maybe I'll try making the L2 cost work for the relative NG. After all I do need to do something with the newfound understanding of Lie algebras.

Hopefully I'll be able to grasp something by the end of the month. Optimization over Lie groups is so mysterious. I must have found it a month ago and I still haven't the damnest clue what the hell it is supposed to be doing."