"10:05am. https://pdfs.semanticscholar.org/b925/43d51c584f47a3e036842b545ee1513c9118.pdf Adaptive On-Line Learning Algorithms for Blind Separation - Maximum Entropy and Minimum Mutual Information
"The learning rule (A) has two important properties: the equivariant property (Cardoso
and Laheld 1996) and the property of keeping W(t) from becoming singular, whereas the
learning rule (14) does not have these properties."

10:35am. Ok, I almost get the stuff in the appendix, but it is so convoluted. What is that inverse towards the end for.

This is classical Amari. I do not get shit of his stuff. But the rule is simple enough and I could use it to try and optimize `sqr(g_inv * g - I)`. I just need to plug in two extra matrix multiplies.

But I want to study this more for a bit. Is there any way for me to simplify this? Because I sure won't be using this for BPRONG as presented."