The bloom is off my love for Dark Souls. I started feeling that it was too easy with a mage so I tried playing it with a thief and then it started absolutely killing me. Towards the end there it started feeling that I was actually getting worse as time went on.
There is no way that I am playing the game at full power. If nothing else, I need to figure out of better ways of priming myself for the parries. I'll stop relying on my eyes so much and make the parry dependent on the delay before the shield is struck.

Well, forget that. Let me watch the videos from the search result I linked to yesterday. Then comes the K-FAC paper.

The truth is at this point I do not even care about the natural gradient all that much as I am tired of it. I am just finishing the loose ends. I want to study the Zap-Q and the Rudder papers plus the stochastic approximation books as that is where the true power lies. It bothers me that the update to the Simulacrum blog was so short, but part of the reason is that I wanted to study those papers as they will be so important to my plans. I did not want to form an opinion before going through the motions properly.

Let me just start. I'll start work on implementing BPRONG tomorrow starting with the kernels for the two Cholesky updates.