"11:40am. https://arxiv.org/pdf/1512.00809.pdf Optimal Whitening and Decorrelation
There are various ways of whitening and they can induce different rotations. I need to research this a bit.

"Interestingly, Cholesky whitening always assumes third place in the rankings, either
behind ZCA and ZCA-cor whitening, or behind PCA and PCA-cor whitening. Moreover,
it is the only approach where by construction one pair (z4, x4) perfectly correlates
between whitened and original data."

Yeah, this paper is exactly what I needed in fact.

"This procedure is also known as Mahalanobis whitening. With Q1 = I it is the unique
sphering method with a symmetric whitening matrix."

This matches what I read in the decorrelated batch norm paper with regards to ZCA.

12:10pm. 8/14. That `z*x` term shows up in differentiable plasticity.

12:20pm. "A consequence of using Cholesky factorization for whitening is that we implicitly
assume an ordering of the variables. This can be useful specifically in time course analysis
to account for auto-correlation (cf. Pourahmadi, 2011, and references therein)."

What does this mean?

12:30pm. https://arxiv.org/abs/1202.1661
Covariance Estimation: The GLM and Regularization Perspectives

Let me take a look at this.

https://arxiv.org/abs/1511.09468
Optimization theory of Hebbian/anti-Hebbian networks for PCA and whitening

Hmmm, this could be interesting.

http://cs229.stanford.edu/proj2010/Zhu-OptimizationBasedWhitening.pdf
Optimization-based Whitening"