"Let me read the two CMA papers that I've been putting off and I will call it a day. Maybe I will start BPRONG tomorrow after all.
https://arxiv.org/abs/1710.03996
A Simple Yet Efficient Rank One Update for Covariance Matrix Adaptation

https://papers.nips.cc/paper/6457-cma-es-with-optimal-covariance-update-and-storage-complexity
CMA-ES with Optimal Covariance Update and Storage Complexity

5:45pm. "Second, the diagonal elements of a triangular Cholesky factor are the square roots of the eigenvalues
of the factorized matrix, that is, we get the eigenvalues of the covariance matrix for free."

Oh, lol. I wrote in the post that a different method would be needed to calculate the eigenvalues. I guess not. This will allow me to do neuron pruning should I want to.

5:55pm. "We also adapted the nontriangular Cholesky factor approach by Suttorp et al. [2009] to the state-of-the art implementation of
the CMA-ES."

Oh, yeah this is something I've been wondering about. So the thing is non-triangular. How strange.

I need to consider this more.

6pm. "Apart from the theoretical guarantees, there are several additional
advantages compared to approaches using a non-triangular Cholesky factorization (e.g., Suttorp et
al. [2009]). First, as only triangular matrices have to be stored, the storage complexity is optimal.
Second, the diagonal elements of a triangular Cholesky factor are the square roots of the eigenvalues
of the factorized matrix, that is, we get the eigenvalues of the covariance matrix for free."

Ah, so the eigenvalue thing only applies to the triangular version. I see. This is great.

I am going to have to put the triangular matrix multiply into the library.

https://docs.nvidia.com/cuda/cublas/index.html#cublas-lt-t-gt-trmm

Here it is. Yeah, I will need this.

"Third, a triangular matrix can be inverted in quadratic time. Thus, we can efficiently compute A^−1t
from At when needed, instead of having two separate quadratic-time updates for A^−1t
and At, which requires more memory and is prone to numerical instabilities."

I need to study the algorithm they introduce to update the Cholesky here.

6:10pm. https://christian-igel.github.io/paper/AMERCMAUfES.pdf
A More Efficient Rank-one Covariance Matrix Update for Evolution Strategies

Here it is. I am really into this. Let me read this and I will call it a day. The plan is starting to form for me."