Had to do some chores. At any rate, I understand enough the Zap Q to implement it. I have no idea how to adapt it to NNs though at this juncture.
One thing I am sure of is - this stuff is genius. I absolutely need to understand it as underneath it lies some kind of gradient normalization scheme. The algorithm is a vision of what lies beyond whitening.

The outer product of the eligibility vector and the new minus the old gradients? Just what is that?