I am in a streange state right now. I am euphoric because all of a sudden I understand the Zap Q algorithm and all the implementation steps, but all those proofs are like a jungle that I rather just cut away. I am really too distracted to focus on them.
As hard as it is, I think Meyn's talk is a lot more approachable than the paper. I should go through it again.

But how about now I move on and start studying the Rudder paper instead? BPRONG, Zap Q and Rudder are the big 3 that I need to understand fully in order to make my plans.

After that I think I'll have enough to start programming. I've been saying this for a while, but I am not going to be studying this forever. I only intend to do it until I am satisfied.