I had to lie down to think a little. I think I won't propagate gradients through the Zap Q component which will be the critic. The reason for that is that I am going to be updating it on each backprop step rather than in the optimization stage. So also propagating the gradients would cause trouble as the whole network would be adapting to weights that have changed which would just add variance. Forget that. Zap Q would be better off as a form of fast memory. I guess that would make it a feature aggregation strategy, but I guess that what I had in mind all along.
Once I saw how much better PG works than Q learning I had already made up my mind to use it. But centering the gradients is a good idea, so it makes sense to do it at some point. And the exact point where it makes the most sense is in the top layer.

That having said, I need to ask the author how a batch version of the algorithm would look like. Let me do that now.