I am really taking my sweet time doing this.
https://stackoverflow.com/questions/51186951/why-does-cublas-not-use-a-1d-array-for-triangular-matrices/51187455#51187455

I was really surprise to see that the level 3 methods do not use a packed format, but it will make testing them a lot easier.

Let me stop here so I can eat and then I will make the tests for the triangular multiply. After that I will put in the triangular solve.