I put in the Cuda functions for the triangular matrix multiply and the triangular solver. I also added the triangular matrix inverse into the mix.
The next step would be to do some real programming. I am going to have to implement the online Cholesky update in a kernel.

This will be a suitable challenge for Spiral.

Let me read that paper again. I need to study the algorithm in depth. I'll probably start work on this tomorrow as it will take me a bit to internalize it.