https://www.ams.org/journals/mcom/1974-28-126/S0025-5718-1974-0343558-6/S0025-5718-1974-0343558-6.pdf Methods for Modifying Matrix Factorizations
The triangual algorithm in the 'A More Efficient Rank-one Covariance Matrix Update for Evolution Strategies' paper is not parallel so now I have to search for one. If that fails, I'll just use the symmetric one instead. I also decided to ask Igel about this. Let me do this for a while and I will call it a day. I am starting to get tired from this.