I finally got everything to work. The Cholesky update is finally operational.
Let me take a short break here and then I will continue.

Forgot to mention it, but it seems A z z^t and z z^t A are not the same thing. That is because the Cholesky factor is not symetric. It completely took me by surprise, but the product of two symetric matrices is not necessarily symmetric.

Wow, I guess I will need that transpose after all. I guess I'll put geam as well into the library to get it. And it seems I put in the symmetric matrix multiply in for nothing.