"2:20pm. It seems that swapping n and m is enough to do the trick. Ok.
2:25pm. The transpose function works now as well.

But that having said, maybe it would be more straightforward not to bother with the transpose. I need to do two transposes as a matter of fact.

2:45pm. https://stackoverflow.com/questions/51223342/would-two-transposes-and-an-operation-be-more-efficient-than-the-same-operation

It is easier to just ask."

Forget the update for a bit.

Let me do the constant factors now. I'll do them separately from the test for the update.