Well, the constant factor kernels are straightforward enough. What is troubling me right now would be how to efficiently do the transpose.
I want to do it without allocating memory.. I know how to do it for square matrices, but I am a bit confused when it comes to non-square ones. If I wanted to do it inplace, I am not exactly sure where I should be putting the elements.

Hmmmm...ah, yeah it can't be done. I can see it now.

I think I should just satisfy myself with an inplace transpose for a square matrix. Ok.

Let me work on this for a bit.

Maybe I might just finish and the two Cholesky updates before the day is out.