I had to rest for a bit, but now let me go at it again. It is time for the final stretch. If possible I will complete the two Cholesky updates and put them into the library. This will allow me to start work on BPRONG tomorrow.
It has been a long time in coming, but I will finally get it. Then I will implement Zap Q. Assuming that is enough, it will allow me to focus on poker for a while. If not I will move to implementing Rudder related stuff.

Well, that is ways off.

For now let me just focus on what is on my plate.