11:25am. I did some cleaning up. I added a note in the `inplace_transpose` that it would be better to iterate in a strided fashion in order to avoid recomputing the indices for large matrices.
11:35am. Right now I am thinking of how am I going to do the centering. I have everything else planned out in my head.

Ah, fuck it. I'll do straightforward averaging. I will do it based on the original mean rather than the iterative one. It feel like a more stable choice somehow for recurrent nets even if it will be a bit more computationally expensive. But well, the biases are not a hot spot anyway so there is no need to worry about this too much.

11:45am. Let me take a break here. Yeah, I know that I did not do much, but planning counts and now that I have it all on hand, I think I'll be finish BPRONG by the time the day ends. Those Cuda Kernels were the really hard parts. What lies ahead is merely putting the pieces together. It won't be a problem.