``` [|     [|18605.85; -497306.1; 131403.1; 124523.8; 957474.7; 130898.3|]     [|-285591.3; 7702774; -2046288; -1921642; -1.483155E+07; -2036315|]     [|75011.19; -2041167; 545075.1; 507388.6; 3930545; 541878.6|]     [|47364.11; -1265340; 334240.7; 316901.6; 2436177; 332975.8|]     [|606155.3; -1.635175E+07; 4344400; 4079037; 3.148506E+07; 4323141|]     [|68483.56; -1862341; 497134.6; 463057; 3586171; 494254.5|] |]
[|
    [|18604.91; -497286.1; 131398.2; 124519; 957435.3; 130893.4|]
    [|-285578.1; 7702500; -2046219; -1921575; -1.483101E+07; -2036247|]
    [|75007.94; -2041098; 545057.6; 507371.9; 3930409; 541861.1|]
    [|47361.88; -1265292; 334229; 316890.1; 2436083; 332964.2|]
    [|606128.4; -1.635119E+07; 4344260; 4078903; 3.148393E+07; 4323001|]
    [|68480.59; -1862277; 497118.5; 463042; 3586045; 494238.7|]
|]
```

The two updates track each other pretty closely so this problem cannot be due to a faulty kernel. Rather the problem is with the update itself. Maybe the paper is wrong. I am wondering why the update in the other paper is so different from this. I might need to ask the author this.

I need to step back a little and go over it with a comb. I cannot possibly get PRONG to work until I can get this to work.