https://arxiv.org/pdf/1303.4622.pdf Constructing numerically stable Kalman filter-based algorithms for gradient-based adaptive filtering
A lot of the Google hits are from the 70s. When I Googled the Potter formula I got a lot of stuff about Harry Potter. Maybe it is time I look for a better search engine? The above papers seems like it will be interesting.

https://www.youtube.com/results?search_query=square+root+filtering&page=&utm_source=opensearch

There are also videos about this on Youtube.

Let me do some studying on this.