I think I need a short break here.
I did a lot of experimenting with different learning rates and the number of steps. This works quite well and hopefully should do so in a real net.

The next step will be to actually make a kernel out of this. I'll get rid of the regular update in that module since thanks to my insight I do not actually need the inverse in order to reproject.

Welll, I already have most what I need in what I made previously, I just need to touch it up a little.