It is not going well, I am going to put in QR and LU factorization into the library tomorrow. I guess I finally have a reason to do this.
Right now I just tried to somehow bias the matrix, but that just made it non-invertible. What I'll do tomorrow is use Q D Q^T with some hard D in order to reproject a random Gausian to some difficult condition number and covariance. I need to get a handle on this.

Today, well I guess I do not feel like Dark Souls that much, not as much as before, but since I am close to being done with the game I might as well go for it.

I did well today. It won't be long until I finish PRONG and get to work on Zap Q. Once all of that is out of the way, I'll have enough to make reliable RL players. If not, I'll start work on Rudder. Everything will go according plan, for as long as I am able to.