Hmmmm, for some reason 4 days of entires at the beginning are missing from my journal. I have no idea what to make of this. Well, forget it.
First comes the morning prep and then comes the programming as usual.

There is something I am confused aboout regarding covariance. I could use the root of the matrix to correlate variables, but what if some of the entries are negative? Does that mean the root of the matrix is invalid?

That can't be it.

My understanding of linear algebra is really garbage.

Maybe I'll ask about it on the math SO after I finish putting in the two decompositions.