Agh, this is going to be such a pain. I can tell just by looking at the API. Putting in the matrix inverse was a pain last time, but it did come quite useful in testing the new algorithm, so I should finish the job here.
Forget hopes and dreams and everything else. I'll dedicate a day or two to figuring out how the two factorization functions work and put them in. Then I will be free from that. And once I finish thoroughly testing the Cholesky stuff, I'll be free of it as well.

Then I will finish PRONG and will be free of that.

I'll eventually get to the point where I'll just be able to focus on RL if I follow the process. Let me get cracking.