I am having problems here. I am not sure how to adapt QR factorization for row major tensors at all. Yeah, I know I could just transpose them, but I am not familiar enough with QR factorization to know what I should do with the result of that afterwards and it is not clear to me what the outputs of geqrf_batched are supposed to be.
...Yeah, I need to transpose the input, then I need to swap the m and n arguments. ...Maybe.

Do I need to transpose the results back?

I have no choice, but to ask on the Cuda SO. If this gets downvoted there, I will steer clear of that place for good.

Edit: Forget about asking. I am reasoning things through. Though my main effort right now is into warming myself up to actually doing this.

Doing all the drudge work of adding all those dependencies and doing the transposes as such is not my idea of a good time.

I have a half mind wanting to just take Mnist, pare it down and experiment on that.

No, I must do it. I have to do it. Those solver functions might become a convenience for something else at some point.

...Let me struggle for a bit more with this. If only Cuda did not have this stupid column major bullshit that I have to deal with.

I am going to ask on the math SO with regards to how the LU and the QR factorizations act when the inputs to them are trnasposed. Do I have to transpose both the input and the output or will maybe just the output do?