Forget that the surge of interest in the language for now. I need to start the programming day rather than dwell on this. Continuing from what I did yesterday, I want to figure out how the batch size affects the symmetry of the inverse factor. I also do not understand why the one step update is not symmetric. It certainly looks like it wants to be. I want to try initializations other than the identity.
Ultimately, I think the exact symmetry of this is not that important. I expect that I could use the symmetric matrix multiply even on the quasi-symmetric matrix and it would still give good results. I am not interested in pushing optimizations at this point though. I'll consider it in the future after I've finished implementing PRONG.

My goal for the day will to finally clear these tests and then move to implementing the PRONG layer.