``` The absolute maximum of A_diff is 0.1072566 with step size 1 The absolute maximum of A_diff is 0.02800446 with step size 2 The absolute maximum of A_diff is 0.01414589 with step size 3 The absolute maximum of A_diff is 0.008319365 with step size 4 The absolute maximum of A_diff is 0.00536552 with step size 5 The absolute maximum of A_diff is 0.003335387 with step size 6 ... The absolute maximum of A_diff is 0.0001472533 with step size 32 ```
As can be seen, just a few steps are needed to bring the difference between the symmetric terms down by an exponential amount.

With the standard batch sizes used in supervised learning the error is negligible. But I do not think in practice the error with k=1 would be an issue either.

Well, with that I am done with this. Let me take a break here. After that I'll start work on PRONG. No, before I move to that I should take a look if the dual updates still track each other even with averaging. I'll do that after the break.