For some reason testing takes surprisingly long. I need a break. I breezed through the documentation for Felix. The approach it takes for C++ interop is suprisingly similar to Spiral, but I do not think the languages have much in common.
I'll be interested in how will they handle the performance issues without inlining guarantees. Right now, I am rough on Julia because of that.

Spiral is a lot simpler language than Felix. This is not great across the board though - the lack of formalism means making tooling for it will be hell if it ever gets popular. It is likely the language will have to be redesigned in some fashion to allow for that.

Anyway, let me save the data here. This is what I sent to Krause. I need to rest a bit here.

"I made an error again. I was wondering why it was blowing up going from 256 to 512 and I just remembered that I forgot to adjust the projection matrix. It was set for the 8x8 matrix which would cause the outputs to 512 to be vastly oversized. Probably if I set it to something sensible it would allow me to use higher learning rates, but I doubt it would change any of the conclusion drawn from the data with respect to the symmetry error.

...It seems that sensible scaling makes no difference at all.  Nevermind.

But going from '100 passes over 1024 Mnist examples randomly projected to a 512 dimensional space with the learning rate is 0.001' to ' 200 passes over 1024 Mnist examples randomly projected to a 512 dimensional space with the learning rate is 0.001' makes the errors much worse when they are bad.

```
1, NaN
2, NaN
3, 1.81971E+14
4, 93.80208
5, 5.656786
6, 4.126465
7, 3.194387
8, 2.427779
9, 2.076558
10, 1.62397
11, 1.428761
12, 1.183509
13, 1.000727
14, 0.8773472
15, 0.754323
16, 0.7081251
17, 0.7037292
18, 0.6489496
19, 0.5869555
20, 0.5971389
21, 0.546948
22, 0.5655413
23, 0.5461698
24, 0.5141549
25, 0.4649711
26, 0.4817905
27, 0.4444346
28, 0.4408298
29, 0.3813033
30, 0.3753654
31, 0.3502119
32, 0.3063558
200 passes over 2048 Mnist examples randomly projected to a 512 dimensional space. Learning rate is 0.001. Float32 types are used for all the matrices.
```

Now k=2 has blown up and k=3 is on the verge of it. I'll do the 5x number of passes next.

```
1, NaN
2, NaN
3, NaN
4, NaN
5, NaN
6, NaN
7, NaN
8, 1.08584E+17
9, 444894.7
10, 9.000236
11, 8.018881
12, 6.897162
13, 5.874614
14, 5.089999
15, 4.557597
16, 3.854511
17, 3.602167
18, 3.132749
19, 2.91012
20, 2.557649
21, 2.281958
22, 2.180998
23, 1.903274
24, 1.87501
25, 1.718912
26, 1.600254
27, 1.457733
28, 1.416787
29, 1.266992
30, 1.214243
31, 1.106944
32, 1.117683
1000 passes over 2048 Mnist examples randomly projected to a 512 dimensional space. Learning rate is 0.001. Float32 types are used for all the matrices.
```

This one took a while. It seems that when the learning rate is set too high, the errors continue to get worse even with averaging. I'd guess with long enough time to run they'd all blow up eventually."

Edit: Forgot to commit this an hour ago. Let me start work on PRONG.