I am looking at how the back whitening is being updated and it seems it is slowly, but surely diverging. The pattern is of diagonals just growing.
Did I make some horrible mistake? Is the data I sent to Krause somehow faulty due to that? Or is it something banal like the data being intrinsically non-invertible?

I'll dedicate tomorrow to get to the bottom of this. Really, I should have expected nonsense like this and did in fact. Was there ever a moment where I got anything right on the first try?