Actually, let me do this work now. I am just so dubious about how the costs for this are all zero.
Edit: Well, the norm is something like 1E-5 and after it is passed through the learning rate it is more like 1E-9. sqrt (1 + 1E-9) comes out to 1 on my computer. So none of the factors ever get updated.

This is unbelievable.

I need to repeat the experiments from today and yesterday and this time look at the damn factor matrix. This is not easy since it is something like 512x512, but I am going to have to do it.

Alternatively, I could just increase the damn learning rate.

Let me try that first.